requirements.txt
openai
tiktoken
python-dotenv
